+++ resources/srp-base/fxmanifest.lua
@@ -1,0 +1,21 @@
+--[[
+    -- Type: Manifest
+    -- Name: fxmanifest
+    -- Use: Declares srp-base resource scripts
+    -- Created: 2025-09-06
+    -- By: VSSVSSN
+--]]
+
+fx_version 'cerulean'
+game 'gta5'
+
+lua54 'yes'
+
+server_scripts {
+    'server/**/*.lua'
+}
+
+shared_scripts {
+    'shared/**/*.lua'
+}
+