+++ resources/srp-base/server/modules/jobs.lua
@@ -1,0 +1,22 @@
+--[[
+    -- Type: Module
+    -- Name: jobs
+    -- Use: Handles job related RPC calls
+    -- Created: 2025-09-06
+    -- By: VSSVSSN
+--]]
+
+local M = {}
+
+--[[
+    -- Type: Function
+    -- Name: handle
+    -- Use: Processes job envelope
+    -- Created: 2025-09-06
+    -- By: VSSVSSN
+--]]
+function M.handle()
+    return nil, "not_implemented"
+end
+
+return M