+++ resources/srp-base/server/modules/world.lua
@@ -1,0 +1,22 @@
+--[[
+    -- Type: Module
+    -- Name: world
+    -- Use: Handles world related RPC calls
+    -- Created: 2025-09-06
+    -- By: VSSVSSN
+--]]
+
+local M = {}
+
+--[[
+    -- Type: Function
+    -- Name: handle
+    -- Use: Processes world envelope
+    -- Created: 2025-09-06
+    -- By: VSSVSSN
+--]]
+function M.handle()
+    return nil, "not_implemented"
+end
+
+return M